requirements.txt
colorlog==6.8.2
h5py==3.12.1
k2eg==0.2.7
lume-model==1.8.1
matplotlib==3.10.1
mlflow==2.12.1
networkx==3.4.2
numpy<2.0.0
p4p==4.1.12
pydantic==2.10.6
pytest==8.3.4
PyYAML==6.0.2
Requests==2.32.3
sympy==1.13.1
networkx
# tensorflow
# torch
# botorch
